chore: rename to menu-select
